requirements.txt
requests~=2.25.0
beautifulsoup4~=4.9.3
python-dateutil~=2.8.1